fix: update the final score to upto 1 decimal place

+++ src/main/java/com/gitrepositoryscore/domain/RepositoryWithPopularityScore.java
@@ -2,10 +2,12 @@ package com.gitrepositoryscore.domain;
 
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 public class RepositoryWithPopularityScore {
     String fullName;
-    long score;
+    BigDecimal score;
     int stars;
     int forks;
     String updatedAt;
